pyproject.toml
[project]
name = "chatbot"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "celery>=5.6.2",
    "chromadb>=1.5.1",
    "fastapi[standard]>=0.133.0",
    "google-genai>=1.65.0",
    "jinja2>=3.1.6",
    "jwt>=1.4.0",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.1",
    "langchain-groq>=1.1.2",
    "langchain-openai>=1.1.10",
    "langchain-text-splitters>=1.1.1",
    "langchain-unstructured>=1.0.1",
    "langgraph>=1.0.9",
    "passlib[bcrypt]>=1.7.4",
    "pip>=26.0.1",
    "psycopg2-binary>=2.9.11",
    "pwdlib[argon2]>=0.3.0",
    "pydantic-settings>=2.13.1",
    "pyjwt>=2.11.0",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "redis>=7.2.0",
    "sqlalchemy>=2.0.47",
    "sse-starlette>=3.3.2",
    "unstructured[all-docs,pdf]>=0.21.5",
    "uvicorn>=0.41.0",
]

[dependency-groups]
dev = [
    "black>=26.1.0",
    "pytest>=9.0.2",
    "ruff>=0.15.2",
]
